+++ src/lib/generateSchemaJson.ts
