+++ src/lib/generateSchemaXml.ts
