+++ src/lib/hasExample.ts
